+++ src/app/services/util.service.ts
@@ -264,5 +264,24 @@ export class UtilService {
     return result;
   }
 
+  distanceInKmBetweenEarthCoordinates(lat1, lon1, lat2, lon2) {
+    console.log(lat1, lon1, lat2, lon2);
+    const earthRadiusKm = 6371;
+
+    const dLat = this.degreesToRadians(lat2 - lat1);
+    const dLon = this.degreesToRadians(lon2 - lon1);
+
+    lat1 = this.degreesToRadians(lat1);
+    lat2 = this.degreesToRadians(lat2);
+
+    const a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+      Math.sin(dLon / 2) * Math.sin(dLon / 2) * Math.cos(lat1) * Math.cos(lat2);
+    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+    return earthRadiusKm * c;
+  }
+
+  degreesToRadians(degrees) {
+    return degrees * Math.PI / 180;
+  }
 
 }